js/svg: stagger all coords at same y-offset
When a testsuite carries multiple runs (separate suites with matching
name), y-coords for shared testcases are only staggered if the testcase
results match - if the results don't match then they'll be put in
separate pass/skip/fail y-buckets so staggering is unnecessary.
This turns out to not be worth it for a couple of reasons:
- we need to track the current y-axis stagger for each testcase
- the line plotted for a suite can be a little hard to follow

Improve this by using a consistent y-axis stagger for all points of
a test run. Drop per-testcase y-axis stagger state.

Signed-off-by: David Disseldorp <ddiss@suse.de>